+++ settings/permissions/PermissionSetDetails.js
@@ -12,7 +12,6 @@ import { Field, reduxForm } from 'redux-form';
 import PermissionSet from './PermissionSet';
 
 class PermissionSetDetails extends Component {
-
   static propTypes = {
     stripes: PropTypes.shape({
       hasPerm: PropTypes.func.isRequired,
@@ -26,6 +25,8 @@ class PermissionSetDetails extends Component {
         PUT: PropTypes.func.isRequired,
       }),
     }),
+    initialValues: PropTypes.object,
+    tellParentTheRecordHasBeenCreated: PropTypes.func,
   };
 
   constructor(props) {
@@ -41,6 +42,8 @@ class PermissionSetDetails extends Component {
     this.connectedPermissionSet = props.stripes.connect(PermissionSet);
 
     this.state = {
+      newSet: !props.initialValues.permissionName,
+      selectedSet: undefined,
       confirmDelete: false,
     };
   }
@@ -53,9 +56,17 @@ class PermissionSetDetails extends Component {
 
   saveSet() {
     const set = this.state.selectedSet;
-    this.props.parentMutator.permissionSets.PUT(Object.assign({}, set, {
-      subPermissions: set.subPermissions.map(p => p.permissionName),
-    }));
+    if (this.state.newSet) {
+      this.props.parentMutator.permissionSets.POST(Object.assign({}, set, {
+        mutable: true
+      }));
+      this.setState({ newSet: false });
+      this.props.tellParentTheRecordHasBeenCreated();
+    } else {
+      this.props.parentMutator.permissionSets.PUT(Object.assign({}, set, {
+        subPermissions: (set.subPermissions || []).map(p => p.permissionName),
+      }));
+    }
   }
 
   beginDelete() {
@@ -88,15 +99,15 @@ class PermissionSetDetails extends Component {
   }
 
   render() {
-    const { selectedSet, stripes } = this.props;
+    const { selectedSet, stripes, initialValues } = this.props;
     const disabled = !stripes.hasPerm('perms.permissions.item.put');
-
     return (
       <Pane paneTitle={`${selectedSet.displayName || 'Untitled'}`} defaultWidth="fill" fluidContentWidth>
         <form>
 
           <section>
-            <h2 style={{ marginTop: '0' }}>About</h2>
+            <h2 style={{ marginTop: '0' }}>{this.state.newSet ? 'Create' : 'Edit'} permission-set</h2>
+            <Field label="Name" name="permissionName" id="permissionName" component={Textfield} required fullWidth rounded validate={this.validateSet} onBlur={this.saveSet} disabled={disabled || initialValues.permissionName} />
             <Field label="Title" name="displayName" id="displayName" component={Textfield} required fullWidth rounded validate={this.validateSet} onBlur={this.saveSet} disabled={disabled} />
             <Field label="Description" name="description" id="permissionset_description" component={TextArea} validate={this.validateSet} onBlur={this.saveSet} fullWidth rounded disabled={disabled} />
           </section>